ddl: remove int precision
This only affects display, not storage, and is not supported by PostgreSQL.

+++ util/src/main/resources/cleanAccount.sql
@@ -3,7 +3,7 @@ DELIMITER //
 CREATE PROCEDURE cleanAccount(p_account_key char(36))
 BEGIN
 
-    DECLARE v_account_record_id int(11) /*! unsigned */;
+    DECLARE v_account_record_id bigint /*! unsigned */;
 
     select record_id from accounts WHERE external_key = p_account_key into v_account_record_id;
